refactor: switch back to app directory
