+++ apps/website/src/app/layout.tsx
@@ -1,3 +1,4 @@
+import 'styles.css';
 import Footer from 'components/Footer';
 import Navbar from 'components/Navbar';
