+++ apps/website/src/components/Command.tsx
@@ -1,3 +1,5 @@
+'use client';
+
 import { useState } from 'react';
 import { HiOutlineClipboardCopy, HiOutlineClipboardCheck } from 'react-icons/hi';
