+++ apps/website/src/components/Navbar.tsx
@@ -1,3 +1,5 @@
+'use client';
+
 import { useState } from 'react';
 import { HiOutlineMenu, HiOutlineX } from 'react-icons/hi';
