--- apps/website/src/pages/_app.tsx
@@ -1,5 +1,0 @@
-import '../styles.css';
-
-export default function MyApp({ Component, pageProps }) {
-	return <Component {...pageProps} />;
-}
